feat(ui): add low performance mode
This commit introduces a new "low performance mode" to the app.
This mode can be toggled from the Admin screen and is designed to
improve performance on older devices by disabling animations and
other performance-intensive features.

A new `PerformanceContext` is created to provide the current
performance mode to all components. The `PulsingCircle` component
is updated to respect this setting.